Moved getNotification to NotificationPublisher
