+++ Habitica/src/com/habitrpg/android/habitica/MainActivity.java
@@ -1,5 +1,8 @@
 package com.habitrpg.android.habitica;
 
+import android.app.AlarmManager;
+import android.app.PendingIntent;
+import android.content.Context;
 import android.content.DialogInterface;
 import android.content.Intent;
 import android.content.SharedPreferences;
@@ -135,6 +138,9 @@ public class MainActivity extends AppCompatActivity implements HabitRPGUserCallb
         this.hostConfig = PrefsActivity.fromContext(this);
         if(!HabiticaApplication.checkUserAuthentication(this, hostConfig))
             return;
+
+        //Check if reminder alarm is set
+        scheduleReminder(this);
 
         HabiticaApplication.ApiHelper = this.mAPIHelper = new APIHelper(hostConfig);
 
@@ -738,5 +744,28 @@ public class MainActivity extends AppCompatActivity implements HabitRPGUserCallb
         }
 
         return super.onKeyUp(keyCode, event);
+    }
+
+    private void scheduleReminder(Context context) {
+
+        SharedPreferences prefs = PreferenceManager.getDefaultSharedPreferences(this);
+        String timeval = prefs.getString("reminder_time", "19:00");
+
+        String[] pieces = timeval.split(":");
+        int hour = Integer.parseInt(pieces[0]);
+        int minute = Integer.parseInt(pieces[1]);
+        Calendar cal = Calendar.getInstance();
+        cal.set(Calendar.HOUR_OF_DAY, hour);
+        cal.set(Calendar.MINUTE, minute);
+        long trigger_time = cal.getTimeInMillis();
+
+        Intent notificationIntent = new Intent(context, NotificationPublisher.class);
+        notificationIntent.putExtra(NotificationPublisher.NOTIFICATION_ID, 1);
+        notificationIntent.putExtra(NotificationPublisher.CHECK_DAILIES, false);
+
+        PendingIntent pendingIntent = PendingIntent.getBroadcast(context, 0, notificationIntent, PendingIntent.FLAG_UPDATE_CURRENT);
+
+        AlarmManager alarmManager = (AlarmManager) context.getSystemService(Context.ALARM_SERVICE);
+        alarmManager.setInexactRepeating(AlarmManager.RTC_WAKEUP, trigger_time, AlarmManager.INTERVAL_DAY, pendingIntent);
     }
 }
